ProjectB/guestbook/src/main/java/com/example/guestbook/repository/LogRepository.java
package com.example.guestbook.repository;

import com.example.guestbook.domain.Log;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
import org.springframework.stereotype.Repository;

import javax.sql.DataSource;
import java.util.HashMap;
import java.util.Map;

@Repository
public class LogRepository {
    private final JdbcTemplate jdbcTemplate;

    @Autowired
    public LogRepository(DataSource dataSource) {
        jdbcTemplate = new JdbcTemplate(dataSource);
    }

    public Log insert(Log log) {
        SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate);
        simpleJdbcInsert.withTableName("log").usingGeneratedKeyColumns("id");

        Map<String, Object> parameters = new HashMap<>();
        parameters.put("ip", log.getIp());
        parameters.put("method", log.getMethod());
        parameters.put("regdate", log.getRegdate());

        Number key = simpleJdbcInsert.executeAndReturnKey(new MapSqlParameterSource(parameters));
        log.setId(key.intValue());
        return log;
    }
}
